[deoptimizer][wasm] Fix DCHECK for Float32 result in lazy deopt of JS-to-Wasm wrapper
The Float32(float) constructor DCHECKs that the provided float is not a
NaN value because for Wasm it is important to not accidentally change
the bit representation of a NaN (as the wasm spec enforces keeping the
bit representation unmodified for certain operations.

For the JS-to-Wasm-wrapper this doesn't matter as JS does not have such
guarantees and different NaN bit patterns should not result in
observable differences in behavior.

Fixed: 431542997
Change-Id: Ieadf5641534981410be6c423c8b4b3ede8ede75d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6732847
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101406}