[gcmole] GCMole should accept forward declarations of scopes
GCMole currently does not support forward declarations of the scopes
used to exempt code. This was not an issue for DisableGCMole scope
since it was never used as a witness. For ConservativePinningScope&,
which can be used as a witness, supporting forward declarations is
needed.

Bug: 431698644
Change-Id: I14fcc13123e6120c2e0b7db4e185ca390439d341
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6732845
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101413}