[heap] Use kSharedOld for all in-place internalizable strings
This CL cleans up the initialization logic for
allocation_type_for_in_place_internalizable_strings_ in Heap. With
--shared-string-table we always want to allocate in-place
internalizable strings in the shared space using
AllocationType::kSharedOld. While this property was already
upheld prior to this change, the logic implementing this seems broken.

So far the implemented logic seemed to imply that we would use
AllocationType::kOld on the shared space isolate but kSharedOld on
client isolates. This was quite confusing because our mental model was
always that strings in the shared string table should always be
shaerable (which means either be in read-only or shared space). However,
as Patrick thankfully found out this wasn't actually the case. This is
because we already loaded is_shared_space_isolate_ before it was set to
true. So we could never observe `is_shared_space_isolate_ == true` and
because of that always used kSharedOld when allocating in-place
internalizable strings after all.

In order to prevent similar issues in the future, this CL adds a
new DCHECK which ensures that we invoke is_shared_space_isolate() only
after its corresponding field was initialized. In addition to that
this CL also DCHECKs that strings in the shared string table always
either point into the read only or shared space. This should make
us more confident in the future that this invariant is upheld.

Change-Id: I0873e61fd5fae52ca4cafa92c35253036c984878
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6732046
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101420}