[cfi] Use VirtualAddressSubspace with PKEY for cfi-related pointer tables
In https://crrev.com/c/6657446 we introduced the ability to set a PKEY
on a VirtualAddressSubspace, which then guarantees that all its pages
are protected by that key. With this CL we now use this mechanism for
our pkey-protected pointer tables. This probably fixes a bug where
previously, the pkey protection was lost for segments that were
decommitted/freed (and then reallocated) as that will replace the page
with a fresh page (without a pkey) under the hood.

Change-Id: I7321ff207a573a4c042a57dc5f8038404719c753
Cq-Include-Trybots: luci.v8.try:v8_linux64_pku_dbg,v8_linux64_pku_rel
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6759891
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101476}