[wasm] Fall back to previous undefined NaN when calling Wasm
The undefined double experiment changes the NaN bit pattern of
undefined, which is an observable difference through wasm when the value
is reinterpret casted to e.g. integral types. To silence fuzzers, we for
now fall back to the old bit pattern when calling into Wasm.

Bug: 428771931,385155404
Change-Id: Ice2f9885658cfe333db67a7dd91ef6591400da71
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6759911
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101448}