Check current isolate when accessing MemoryChunk metadata
With the sandbox enabled, the attacker may corrupt the page metadata
index to point to a metadata of the page belonging to a different
isolate. If that different isolate gets destroyed and the memory chunk
is not pulled, then the main isolate may get UaF and hence escape the
sandbox.

The CL embeds the owner Isolate into the MetadataTable so that whenever
a metadata is accessed, the caller would sandbox-check the current
isolate against the owner isolate.

Bug: 430498032
Change-Id: Iea8b2ab1e062cacb8d4ec3feb08738bfbe493baa
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6722280
Auto-Submit: Anton Bikineev <bikineev@chromium.org>
Commit-Queue: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101453}