Add a flag to control the ICU datetime format workaround
The purpose of this flag is to support 3 behaviors:

- Unconditionally replacing ASCII space with NNBSP (current behavior)
- Only doing it for English locales (likely web compatible)
- Never doing the replacement (likely to break some web sites)

An experiment will be run in Chrome to limit the workaround to English
only. If that is successful we may also consider the final step and
removing the workaround, again in a controlled experiment.

Bug: 432742367
Change-Id: I349cced7de1fe5d6ff16841d719325c76c374e19
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6771125
Reviewed-by: Patrick Thier <pthier@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Philip Jägenstedt <foolip@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101548}