[wasm][fuzzer] Add random module generation fuzztest
This adds the equivalent of the "wasm_compile_fuzzer" as a fuzztest.

It extracts the `SyncCompileAndExecuteAgainstReference` method out of
`WasmExecutionFuzzer::FuzzWasmModule` in order to be able to reuse it in
the FuzzTest.

In contrast to the libfuzzer fuzzers, the first byte of the input is not
used as a "configuration byte". Instead, the configuration (tier mask
and debug mask) is passed explicitly.

Also, the static setup (setting experimental flags) is done in the
FuzzTest constructor instead of dynamically when executing the first
testcase.

R=mliedtke@chromium.org

Bug: 433343732
Change-Id: I9252417a6a1cb4971feb233da833611c1d33d504
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6771338
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101573}