[temporal] Implement Ch15 of Temporal
https://tc39.es/proposal-temporal/#sec-temporal-intl
Use 12 unused bits from the flags to remember which option was present in the optinos during the constructor so we can lazy create the SimpleDateFormat for Temporal during the format time. This avoid memory penalty and performance penalty on the Intl.DateTimeFormat when it is not used with Temporal (such as in (new Date()).toLocaleString(). If we resolve all format during constructor, we may cause performance problem and blow the memory for old (and most common) use cases.


Bug: 401065166
Change-Id: I70b9f0d37a1cfbf00d80e382b594a6e7829c6fd3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6770445
Commit-Queue: Frank Tang <ftang@chromium.org>
Reviewed-by: Manish Goregaokar <manishearth@google.com>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101584}