Remove CSA::LoadJSFunctionCode
With leap tiering (which is the default) this method loads the dispatch
handle from the JSFunction and then reads the code from the dispatch
table. This can lead to hard-to-spot double reads of the dispatch
handle.

Thus remove that method.

The one caller now explicitly reads the `code` field from the
`JSFunction` if leaptiering is disabled, and reads from the dispatch
handle otherwise.

R=olivf@chromium.org

Bug: 430960844
Change-Id: I13ef655bfe3772177db59f1ad8ca5907bec24386
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6771793
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101592}