[GC] Move consumed memory metrics to StartCycle
Current and limit consumed memory can change between the
start of incremental marking and the start of atomic pause.
Since the delta between the 2 isn't very actionable,
this CL captures the values in StartCycle, which is at
incremental marking start for a cycle that has incremental marking.

Bug: 373759996
Change-Id: Ice295b263bbe42d44bd820edceed25184bffde27
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6779024
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101599}