[wasm] Fix sandbox escape via manipulated WasmSuspendingObject
The `callable` stored in the `WasmSuspendingObject` must not be a Wasm
function. This is enforced in the constructor, but if this is changed
later via sandbox corruption it can lead to internal inconsistencies if
we process the import as both a suspending import and an import of an
exported Wasm function.

R=thibaudm@chromium.org
CC=​jkummerow@chromium.org

Bug: 432289371
Change-Id: I2ca6edd51ad93e233d8c8c9757057386956abb24
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6781095
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101706}