[wasm-custom-desc] Optimize prototypes setup in Liftoff
This teaches the LiftoffCompiler the required pattern matching
to detect the typical prototypes setup sequence in the start function,
allowing it to call the optimized builtin instead.
We keep Liftoff code cacheable across all instantiations by emitting
a dynamic check for applicability of the fast path for the current
instance.

Bug: 403372470
Change-Id: Ib79b406d4144ec4a2b8ef3017be3a24d73650dca
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6779123
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101617}