Rename LazyBuiltinsAssembler to JSTrampolineAssembler
This assembler is not only used for lazy compilation builtins any more,
hence rename it to make its intended use more clear.

Also add a short comment to the class and rename the methods to be in
line with other methods on the CodeStubAssembler.

R=olivf@chromium.org

Bug: 430960844
Change-Id: I00ce124cf75e022bae5907ccffbe5170470de6f9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6780590
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101633}