Add flags for parser ablation
Add support to add configurable increase in preparse, parse and bytecode
compile time. The goal is to support ablation studies for investigating
the impact of faster parsing and bytecode generation.

The ablation works using a timer and a busy wait to pass time.
Alternatives considered are running the parser multiple times on the
same input. The problem however is that it is tricky to reset the
state such that the comparison would make sense. For instance we would
have to ensure that the zone memory is freed, string tables reset, and
so on. The trade-off is that we cannot run the ablation on
configurations where we lack a high-precision timer.

Bug: 430201172
Change-Id: I386b18d97a218145641f142f5200ec2632cf7e96
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6768831
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101640}