[fuzzilli] Update crash tests
We have a few crash tests that currently only work if ASan is enabled.
That is bad because we might use them in a configuration that doesn't
have ASan enabled, leading to fuzzer startup failures. As such, we now
enforce that these tests also crash on non-ASan builds.

Bug: 429859158
Change-Id: I2aa2119848e59556f145447ee6ddb89caf6d5418
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6799385
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101686}