[asm] Use symbol for storing single exported function
Do not store the single exported function as a property on the instance
object where it can be retrieved and manipulated easily from user code.
Use a symbol instead which is much less visible and can only be
manipulated using in-sandbox manipulation primitives.

This also removes a convenient hook for manipulating in-sandbox objects
after instantiation but before running the start function.

R=jkummerow@chromium.org

Bug: 430960844, 433809112
Change-Id: I613f270909c70ca9e5214a8c47d8fad870c78a6f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6799599
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101705}