[heap] Fix incremental marking step without atomic field writes
Disabling atomic field writes is used in conjunction with disabling
concurremtn marking to improve TSAN coverage. A safepoint is used to
stop background threads otherwise an incremental step appears like
concurrent marking and leads to false positive TSAN alerts (see comment
in the code).

This was assuming no shared heap. If a shared heap exists, the safepoint
may result in a shared GC, at a time where the heap is not iterable
(memory for an object was allocated but the object hasn't been
initialized yet and doesn't have a map).

To fix while maintaining as much bot coverage as possible, we set up a
safepoint only if it's possible to avoid a global safepoint. If it's not
possible, bail out of the incremental step to avoid false positives.
Avoiding a global safepoint is possible if there is not shared space or
if it's possible to lock the mutex.

Bug: 433940747
Change-Id: Ief1b8f4a0a321c283edef46102a40a6c788dc9ea
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6800225
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101774}