[maglev] Fix DCHECK failure in truncation pass for ReturnedValue
ReturnedValue has a single input, but is is not a FixedInputValueNode,
It is a special node that can accept any input type.
This was causing a DCHECK failure in PropagateTruncationProcessor.

This patch fixes the issue by special-casing ReturnedValue in the
DCHECK. It also adds a comment to ReturnedValue explaining why it's not
a FixedInputValueNode.

Fixed: 435050388
Change-Id: I539a33ef3b3a996083f84c92cc76452d8f55185b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6798370
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101690}