[promise] Add CHECK against empty handle
When resolving a promise, embedders are not allowed to pass us an
empty handle for the value. We have crash reports suggesting that
they sometimes do that. This CL adds a CHECK to prove or disprove
that hypothesis.

Bug: 434637746
Change-Id: Ic9f1e9c6dca0921832a126d501f82804031f12ef
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6802325
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101709}