[macros] Remove some footguns from FOR_WITH_HANDLE_SCOPE
The FOR_WITH_HANDLE_SCOPE defines two nested loops, to allow some code
to run with a periodically resetting HandleScope without
constructing/desctructing that HandleScope on each iteration.

This approach had some footguns -- one is that the whole loop body is a
macro argument (which caused issues in the past with commas inside the
body), but more importantly, a `break` statement inside the inner loop
would not break out of the outer loop. We have been careful not to have
breaks inside this loop that don't also make the limit_check fail in the
outer loop, but this is fragile.

We can use some additional state tracking to make `break` not be broken,
using a "bool in the outer loop" pattern which clang tends to be good at
optimising into gotos.

We can also make the macro support being used as if it were a real for
loop, with the loop body outside the macro rather than inside it, by
using some trickery to get scoped variables without requiring a closing
brace after the body -- then the inner `for` loop can be the last line
of the macro, allowing a (braced) statement to follow it. This trickery
can be encapsulated in another macro (yay macros), and we can re-use it
in Turboshaft's macros which do similar trickery.

This doesn't make the macro any more readable, arguably less so, but
use-sites should be more readable.

Bug: 435184569
Change-Id: I5876928a009787e1da75cba0f6629bb929a65b58
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6802281
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101710}