[hole] Remove hole raw number value
Holes should, in general, not flow through code as values that need to
be read. In particular, we should never perform a JS ToNumber operation
on a Hole, since holes are not JS-visible values.

Historically, holes were considered a form of oddball, and oddballs
_did_ have ToNumber handling (by reading their raw number offset as if
they were a HeapNumber), and in a few cases we ended up being sloppy and
allowed Holes to go through the same path (to read the hole NaN value on
array initialization or elements kinds transitions).

However, in such cases, we _know_ that we are looking for the hole, and
we _know_ that we actually want the hole NaN. So, we can be explicit
about this, with the benefit that we can avoid the hole constant load
and read from its raw number field, and instead can embed the hole NaN
float64 constant where needed.

This required a few changes to Float64Constant encoding in TurboFan
(using an integer bit encoding rather than a double, to avoid losing the
special NaN bits), and a few adjustments to representation changing
which are explicit that we are converting a NumberOrHole rather than
NumberOrOddball when changing a HOLEY_SMI_ELEMENTS array to
HOLEY_DOUBLE_ELEMENTS.

With these changes, we can remove the raw number field from Holes
entirely.

Bug: 434179415
Change-Id: I4d158466dcdcad9f676e6ba7dc1d6cb6e817d440
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6786712
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101712}