[wasm] Forbid recreating WasmInternalFunction for Wasm imports
From the information we have in the importing instance we cannot
correctly recreate the WasmInternalFunction because we would need to
get the declared function index (we only have the import index).
We usually also do not need to recreate them because when importing an
exported Wasm function we cache the func refs. So this can only be
exploited via in-sandbox corruption.

This CL introduces an SBXCHECK to crash in case we try to recreate the
WasmInternalFunction for imported Wasm functions.

R=jkummerow@chromium.org

Bug: 432289371
Change-Id: Iebdc55bb390ae4a937e8bcb8433952a1f31a8383
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6812255
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101725}