[wasm][test] Allow mixing old and new EH
The new fuzztest does not set `--fuzzing` like the libfuzzer targets,
hence it didn't allow mixing the old and new exception handling
proposals so far.

Instead of generally setting the `--fuzzing` flag, we just set this one
specific flag for now. If it turns out that we need more of what
`--fuzzing` implies and changes, we can still switch to that flag in the
future.

R=thibaudm@chromium.org

Bug: 435058040, 433565191
Change-Id: Ia587b86d88dd989028c9419003e666e31108b8a2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6811601
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101728}