[riscv] Fix alignment of jump table entries in assembler tests
There was a good deal of confusion about the necessary alignment of the
individual entries in the jump tables. On 32-bit RISC-V, we typically do
not need to add nop padding. On 64-bit RISC-V, we sometimes use nops to
ensure an even number of instructions so a preestablished alignment is
correctly preserved.

This fixes a sporadic failing DCHECK where we would sometimes generate
an unforeseen nop in a trampoline pool block scope as part of 8 byte
aligning inside the scope.

Change-Id: I8bb9d9bfb9e5ca3506ad37b46ea6a0d55f4ccfb9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6813493
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Ji Qiu <qiuji@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#101843}