[wasm][sandbox] Check sandbox-safety of WasmSuspendingObject
The tentative fix at:
https://chromium-review.googlesource.com/c/v8/v8/+/6781095
Turns out to be incorrect because we don't handle special JS import
call kinds correctly anymore.
Instead, resolve the callable regardless of WasmSuspendingObject, and
check at the end that it is not kWasmToWasm.
A check was already added to catch a more general case of this issue:
https://chromium-review.googlesource.com/c/v8/v8/+/6812255
So this just consolidates the protection for this more specific case.

R=jkummerow@chromium.org

Fixed: 435489665,432289371
Change-Id: Ia979fef5502a4187dc3033df0c09ee437b2917a7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6818338
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101752}