[maglev] Use node types to determine string ops in binary-add
This is similar to TF. Feedback alone is not sufficient because it
describes the types of *both* operands; but for string-add, it is
sufficient for one operand to be a string.

For now, this doesn't help much. It turns the generic Add into
StringConcat or even NewConsString; but in most cases we still have to
call ToString on one of the operands (because range analysis cannot yet
tell us that we can use a faster version of ToString), so we don't save
a builtin call.

However, once range analysis improves, this will automatically generate
better code.

Also note that we need to be careful when either operand is not a
primitive, i.e. it is a JSReceiver. In that case we delegate to
builtins for proper conversion wrt
https://tc39.es/ecma262/#sec-applystringornumericbinaryoperator.

Change-Id: I9ce5e1e426ba1bb59bf92c41b01d6dc1b0dec74b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6797118
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Auto-Submit: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101769}