[wasm][jspi] Throw a SuspendError before calling the import
The JSPI spec says:

> 5. If the entry for async_context in map is not active, then:
>    1. Perform throw a JavaScript exception with a SuspendError exception.
> 6. Prepare to run script with relevant settings.

https://webassembly.github.io/js-promise-integration/js-api/#create-a-suspending-function

In particular, we should throw the SuspendError before we call the
JS import function.

I updated test/mjsunit/wasm/stack-switching-no-stress.js to test that
we don't call the suspending function, and reordered the checks in
wrappers.cc so that we first check whether to throw a SuspendError and
then call the function.

Bug: 42202153
Change-Id: Idcbd87334e6c4cbe7a4b69666c986e746d20e056
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6800062
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101807}