[heap] Add DCHECK to SemiSpaceObjectIterator::Next
In case of a corrupted object size, size could be negative, leading it
either an infinite loop or underflowing the page. This CL adds a DCHECK
that the current object address is monotonically increasing.

A negative size should not lead to a sandbox escape in the revised
implementation.

Bug: 433800618
Change-Id: I85df16827064c7db469f89a7417fc0d435160924
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6827121
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Auto-Submit: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101778}