[api] Introduce EmbedderDataSlot::ToAlignedPointer with type tag
The existing `ToAlignedPointer` uses a generic tag for all types of
content, which is a weakness in the V8 sandbox.

This CL works towards using type specific tags for the values in the
EmbedderDataArray. It deprecates the existing `ToAlignedPointer`, and
introduces new versions, one that accepts a type tag as a parameter, and
another one that is generic and uses a tag range. The tag range that is
allowed to be used for pointers in the EmbedderDataArray is also defined
in this CL.

Bug: 433909571
Change-Id: I545fc21ecc0535cca98586b9f03626eafc4e1fa7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6819929
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101787}