[riscv] Restore status registers after tests
The assembler tests were changing the status registers but not restoring
them. On real hardware this would lead to crashes in the test tear-down
which relied on specific values in the status register.

For example, a test would change the rounding mode, and the tear down
would then get unexpected results.

Change-Id: I6aa37b31f1efa4647dc440f4605e93fabd0ff07b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6827265
Reviewed-by: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Ji Qiu <qiuji@iscas.ac.cn>
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Cr-Commit-Position: refs/heads/main@{#101796}