[maglev] Avoid using TrySmiTagInt32 with 32-bit smis
Before this change, we would sometimes use the try-variant for smi
tagging, but the condition flags wouldn't actually be set even though a
condition code was returned.

This only changes emitted code on RISC-V, where we now avoid setting the
dedicated flags register to zero when we know we do not want to branch
based on its value.

Change-Id: I75dc6dd92344bb1137aff159e16d6a134b5dab61
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6825963
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Cr-Commit-Position: refs/heads/main@{#101783}