[regexp] Store whether a match can be zero length in IrRegExpData
Add bit_field to IrRegExp data and use it to store whether a match can
be zero length or not.
This information is currently stored in the AST. We need to persist it
for assembling from bytecode, as the AST isn't available at code-gen
time.
If we know that a match cannot be zero length, we emit better code for
global RegExps.

Drive-by: Fix IrRegExpData object printing.

Bug: 437003349
Change-Id: Ied9f3eb34a7400993fc5aa587b0b656e957d5377
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6827424
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101927}