perf: ObjectDefineProperty shouldn't always invalidate ArrayIteratorProtector
In `DefineOwnPropertyIgnoreAttributes`, the engine currently calls
`UpdateProtector` to track property changes. However, this is not always
necessary—for example, when a property is changed from writable to
read-only, the protector may not need to be invalidated.

This change refines the logic to prevent unnecessary calls to
`UpdateProtector` in such cases. Specifically, it focuses on
`Array.prototype[Symbol.iterator]`, allowing this property to be made
read-only or non-configurable or enumerable without incurring any
performance penalty.

Change-Id: I52e045c446b77f250717ecbe51eb84fb30552216
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6786371
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101806}