[handles] Refactor Maybe(Direct)Handle implementation
Direct handles must be allocated on the stack. For objects of type
DirectHandle, this is enforced with a slow DCHECK (both for real and
"fake" direct handles, i.e. regardless of the value of the flag
v8_enable_direct_handle). For objects of type MaybeDirectHandle, this
DCHECK was missing.

This CL adds the missing check, both for real and "fake" such handles.
It further refactors the implementation of Maybe(Direct)Handle, so
that they inherit from (Direct)HandleBase, just as the (Direct)Handle
counterparts.

Bug: 42203211
Change-Id: Ie632da0eaddeb53908ca5b05b1824c37203c4b9d
Cq-Include-Trybots: luci.v8.try:v8_linux64_css_dbg
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6832293
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101808}