[wasm] Templatize Wasm (wrapper) graph builder
This is a first CL for the implementation of the inlining of JS-to-Wasm
wrappers with Turboshaft.

The final implementation will reuse the logic to build wrappers with
Turboshaft that already exists in wasm/wrappers.cc, but the existing
class WasmGraphBuilderBas` assumes that several types are templatized on
a fixed set of types:

    using Assembler = compiler::turboshaft::TSAssembler<compiler::turboshaft::SelectLoweringReducer, compiler::turboshaft::DataViewLoweringReducer, compiler::turboshaft::VariableReducer>

In order to implement inlining we want to templatize the same code on
different types Assembler<ReducerList>, used by WasmInJSInliningReducer.

This requires making classes WasmGraphBuilderBase and
WasmWrapperTSGraphBuilder now template classes, and moving a lot of code
from files .cc to files .h.

Bug: 353475584
Change-Id: I6032e857e6eb13e5d56c2f21bd2231de4a3e4dca
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6827999
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#101813}