[test] Remove TEST_FILE_NAME machinery
This was added in 2009 (https://codereview.chromium.org/155161) but is
only used by a single test.
This test can just hard-code its own name in the source.

This is a prerequisite for relanding https://crrev.com/c/6781191 (we
need to be able to find all accessed files from the command-line, and
adding special support for filenames embedded in an argument complicates
things unnecessarily).

R=jkummerow@chromium.org
CC=machenbach@chromium.org

Change-Id: I5526cdceaca8742a545b9ee322749980ad167861
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6829330
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101832}