[wasm] add tracing support for global variables in Wasm
Add support for tracing global variables, similar to the already
existing tracing for reads and writes to Wasm linear memory. This will
be used for detecting and reporting differences in global variables
between the two runs in the differential fuzzing of the Wasm libfuzzers.

Bug: 436219201
Change-Id: Ifabfe4dfdd4a8ac24c4d7c8a1c2705e0bd117c8a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6827702
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Nikolas Kaipel <nikolaskaipel@google.com>
Cr-Commit-Position: refs/heads/main@{#101853}