[api] Add type tag parameters to apis writing EmbedderData
We are working on replacing the generic type tag
kEmbedderDataSlotPayloadTag with type-specific tags. With this CL, API
methods that implicitly used kEmbedderDataSlotPayloadTag get a type tag.

A new enum gets introduced in the API for the type tag parameter, so
that the internal ExternalPointerTag enum does not get exposed. The
embedder is expected to introduce their own enum as well, where the
embedder type tags get organized. The embedder enum can then be cast to
the API enum.

Bug: 433909571
Change-Id: Ib74b57c10d3cf7b49b4e0e4f23b0b77e46a4fa23
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6836349
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101855}