[riscv][wasm] Reverse order of spilled parameters in frame setup
This fixes an issue where we would crash in GC when allocating the
feedback vector, because we would look for the reference parameters in
the wrong stack slots.

The mjsunit/wasm/gc-casts-subtypes test has been flaky because of this
issue for quite a while.

Change-Id: Ib691bfb3528675942c8abd5e77fd5cf0ce9fdbe7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6839045
Reviewed-by: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Commit-Queue: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#101888}