[compiler] Refactor float constant folding in Turboshaft
This patch refactors the constant folding for unary and binary float
operations in the Turboshaft machine optimization reducer.

The result of the constant-folded operation is now stored in a local
variable, and the handling of potential `NaN` results and the final
return is done in a single place after the switch.

This simplifies control flow and reduces code duplication.

R=jkummerow@chromium.org

Bug: 431593798
Change-Id: I83ffeb74fb956aed0304ed21dfb4c543ce858436
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6840797
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101902}