[riscv] Allow trampoline blocking to generate code in jump table tests
The values - and the addresses of the heap-allocated numbers - differ
from run to run, so the amount of code we emit before we block the
trampoline pool isn't fully controlled. This means that we may end up
having to emit the trampoline pool as part of blocking it.

This fixes the flakiness of test-assembler-riscv64/jump_tables3 that we
have seen on the build bots.

Change-Id: I90f5c60dedf7f4a0baf5893f600c9a4f05b6a93b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6842298
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#101866}