[heap] Add CFI directives to PushAllRegistersAndIterateStack on x64
This CL should help backtrace() create proper stack traces
involving PushAllRegistersAndIterateStack. I could verify that this
fixes the stack trace in a local test but also on a ChromeOS bot.

As a potential improvement we could remove the `.hidden` directive
in certain build configurations for PushAllRegistersAndIterateStack.
With this line enabled backtrace() is not able to report the
function name for PushAllRegistersAndIterateStack.

Bug: 437096305
Change-Id: I219e755ee14ce4163d8ea5622d7a23c3047ab0e3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6846039
Reviewed-by: Nikolaos Papaspyrou <nikolaos@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101879}