[turboshaft] Fix direct call in dead code
The pattern `Call(CheckClosure(not-a-closure))` is translated to
`DeoptimizeIf(IsFunctionType(not-a-closure)); Call(not-a-closure)` in
the MachineLoweringReducer. Subsequently we end up lowering this to
a kArchCallJSFunction which assumes the input to be a JSFunction. As
long as we emit instructions which assume the argument to be a
JSFunction nothing happens as the code is dead anyway. However, since
adding a fast-path for known call targets in
https://crrev.com/c/6835869 the back-end assumes the immediate to be a
JSFunction too.

Fixed: 438523769
Change-Id: I66270e9a45b6e5422caff6372fdf5840037f844c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6846049
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101893}