[logging] Fix --log-deopt to not enable deopt tracing
When the `--log-deopt` flag was introduced
(see https://chromium-review.googlesource.com/c/v8/v8/+/2602461),
it was mistakenly used in the `Deoptimizer` constructor as a condition
for creating the trace scope. That scope should instead be controlled
exclusively by the `--trace-deopt` flag.

As a result, enabling `--log-deopt` also produced deoptimization trace
messages on stdout, which is undesirable. These messages cause output
bloat, interfere with stdout parsing, and cannot be suppressed
with `--no-trace-deopt`.

This CL removes `--log-deopt` as a condition for creating the trace
scope.

Bug: 439385225
Change-Id: Ib0a6d31e182c45c1bba7ed402676d1ff51d42dc7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6850335
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101998}